+++ app/src/main/java/eu/kanade/tachiyomi/data/track/mdlist/MdList.kt
@@ -20,7 +20,7 @@ import tachiyomi.source.model.MangaInfo
 import uy.kohesive.injekt.Injekt
 import uy.kohesive.injekt.api.get
 
-class MdList(private val context: Context, id: Int) : TrackService(id) {
+class MdList(private val context: Context, id: Long) : TrackService(id) {
 
     private val mdex by lazy { MdUtil.getEnabledMangaDex(Injekt.get()) }
