+++ app/src/main/java/eu/kanade/tachiyomi/ui/library/LibrarySettingsSheet.kt
@@ -97,7 +97,7 @@ class LibrarySettingsSheet(
             private val unread = Item.TriStateGroup(R.string.action_filter_unread, this)
             private val started = Item.TriStateGroup(R.string.action_filter_started, this)
             private val completed = Item.TriStateGroup(R.string.completed, this)
-            private val trackFilters: Map<Int, Item.TriStateGroup>
+            private val trackFilters: Map<Long, Item.TriStateGroup>
 
             // SY -->
             private val lewd = Item.TriStateGroup(R.string.lewd, this)
@@ -137,7 +137,7 @@ class LibrarySettingsSheet(
                 completed.state = preferences.filterCompleted().get()
 
                 trackFilters.forEach { trackFilter ->
-                    trackFilter.value.state = preferences.filterTracking(trackFilter.key).get()
+                    trackFilter.value.state = preferences.filterTracking(trackFilter.key.toInt()).get()
                 }
 
                 // SY -->
@@ -165,7 +165,7 @@ class LibrarySettingsSheet(
                     else -> {
                         trackFilters.forEach { trackFilter ->
                             if (trackFilter.value == item) {
-                                preferences.filterTracking(trackFilter.key).set(newState)
+                                preferences.filterTracking(trackFilter.key.toInt()).set(newState)
                             }
                         }
                     }
