+++ app/src/main/java/eu/kanade/tachiyomi/ui/manga/MangaPresenter.kt
@@ -1078,7 +1078,7 @@ class MangaPresenter(
                 .map { tracks ->
                     val dbTracks = tracks.map { it.toDbTrack() }
                     loggedServices.map { service ->
-                        TrackItem(dbTracks.find { it.sync_id == service.id }, service)
+                        TrackItem(dbTracks.find { it.sync_id.toLong() == service.id }, service)
                     }
                 }
                 // SY -->
